Move funding helper logic to redux actions
